Tweak

+++ src/main/java/diamondwalker/sscary/randomevent/extrarare/PlayerKillerEvent.java
@@ -1,5 +1,6 @@
 package diamondwalker.sscary.randomevent.extrarare;
 
+import diamondwalker.sscary.data.server.WorldData;
 import diamondwalker.sscary.randomevent.EnumEventRarity;
 import diamondwalker.sscary.randomevent.RandomEvent;
 import diamondwalker.sscary.util.ChatUtil;
@@ -11,16 +12,26 @@ import net.minecraft.util.RandomSource;
 public class PlayerKillerEvent extends RandomEvent {
     @Override
     public boolean execute(MinecraftServer server, ServerPlayer[] validPlayers) {
-        RandomSource random = server.overworld().getRandom();
-        String playerName = validPlayers[random.nextInt(validPlayers.length)].getDisplayName().getString();
-        server.getPlayerList().broadcastSystemMessage(
-                ChatUtil.getEntityChatMessage(
-                        "The" + playerName + "Killer",
-                        "I am going to kill " + playerName + "."
-                ),
-                false
-        );
-        return true;
+        if (validPlayers.length > 0) {
+            RandomSource random = server.overworld().getRandom();
+            ServerPlayer player = validPlayers[random.nextInt(validPlayers.length)];
+
+            WorldData data = WorldData.get(server);
+            if (!data.playerKiller.hasPlayerBeenThreatened(player)) {
+                String playerName = validPlayers[random.nextInt(validPlayers.length)].getDisplayName().getString();
+                server.getPlayerList().broadcastSystemMessage(
+                        ChatUtil.getEntityChatMessage(
+                                "The" + playerName + "Killer",
+                                "I am going to kill " + playerName + "."
+                        ),
+                        false
+                );
+                data.playerKiller.addThreatenedPlayer(player);
+                return true;
+            }
+        }
+
+        return false;
     }
 
     @Override
